refactor: move look up company location into notion
