+++ internal/vacancy-worker/driven/resolver/hqloc/locator.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"net/url"
 
+	"github.com/ghazlabs/idn-remote-entry/internal/vacancy-worker/core"
 	"github.com/ghazlabs/idn-remote-entry/internal/vacancy-worker/driven/resolver/util"
-	"github.com/go-resty/resty/v2"
 	"github.com/openai/openai-go"
 	"gopkg.in/validator.v2"
 )
@@ -26,14 +26,12 @@ func NewLocator(cfg LocatorConfig) (*Locator, error) {
 }
 
 type LocatorConfig struct {
-	HttpClient    *resty.Client  `validate:"nonnil"`
-	DatabaseID    string         `validate:"nonzero"`
-	NotionToken   string         `validate:"nonzero"`
+	Storage       core.Storage   `validate:"nonnil"`
 	OpenaAiClient *openai.Client `validate:"nonnil"`
 }
 
 func (l *Locator) Locate(ctx context.Context, companyName string) (string, error) {
-	companyLoc, err := l.lookupToNotion(ctx, companyName)
+	companyLoc, err := l.Storage.LookupCompanyLocation(ctx, companyName)
 	if err != nil {
 		return "", fmt.Errorf("failed to lookup company location from Notion: %w", err)
 	}
@@ -42,40 +40,6 @@ func (l *Locator) Locate(ctx context.Context, companyName string) (string, error
 	}
 
 	return l.lookupToWeb(ctx, companyName)
-}
-
-func (l *Locator) lookupToNotion(ctx context.Context, companyName string) (string, error) {
-	var respBody NotionResponse
-	resp, err := l.HttpClient.R().
-		SetContext(ctx).
-		SetHeader("Authorization", "Bearer "+l.NotionToken).
-		SetHeader("Content-Type", "application/json").
-		SetHeader("Notion-Version", "2022-06-28").
-		SetBody(map[string]interface{}{
-			"filter": map[string]interface{}{
-				"property": "Company Name",
-				"rich_text": map[string]string{
-					"equals": companyName,
-				},
-			},
-			"sorts": []map[string]string{
-				{
-					"property":  "Last edited time",
-					"direction": "descending",
-				},
-			},
-			"page_size": 1,
-		}).
-		SetResult(&respBody).
-		Post(fmt.Sprintf("https://api.notion.com/v1/databases/%s/query", l.DatabaseID))
-	if err != nil {
-		return "", fmt.Errorf("failed to call notion api to lookup company location: %w", err)
-	}
-	if resp.IsError() {
-		return "", fmt.Errorf("failed to lookup company location from Notion: %s", resp.String())
-	}
-
-	return respBody.GetCompanyLocation(), nil
 }
 
 func (l *Locator) lookupToWeb(ctx context.Context, companyName string) (string, error) {
